pyproject.toml
[project]
name = "rizon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "django-click>=2.5.0",
    "django-extensions>=4.1",
    "django-llm-chat",
    "django-ninja>=1.5.3",
    "django-types>=0.23.0",
    "model-bakery>=1.23.3",
    "ptpython>=3.0.32",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "requests>=2.32.5",
    "rich>=14.3.3",
]

[dependency-groups]
dev = [
    "django-stubs>=5.2.9",
    "flake8>=7.3.0",
    "model-bakery>=1.23.3",
    "pyright>=1.1.408",
    "pytest>=9.0.2",
    "pytest-django>=4.12.0",
    "ruff>=0.15.2",
]

[tool.uv.sources]
django-llm-chat = { git = "https://github.com/sadeh-congnition/django-llm-chat" }
